--- pet/src/main/java/com/pethistory/pet/mapper/UsuarioGettMapper.java
@@ -1,9 +1,0 @@
-package com.pethistory.pet.mapper;
-
-import com.pethistory.pet.dtos.DtoUsuarioGett;
-import com.pethistory.pet.models.Usuario;
-
-public interface UsuarioGettMapper {
-    Usuario toUsuario (DtoUsuarioGett dtoUsuarioGett);
-    DtoUsuarioGett toDtoUsuarioGett (Usuario usuario);
-}
