+++ pet/src/main/java/com/pethistory/pet/mapper/UsuarioMapper.java
@@ -1,9 +1,11 @@
 package com.pethistory.pet.mapper;
 
 import com.pethistory.pet.dtos.DtoUsuario;
+import com.pethistory.pet.dtos.DtoUsuarioGett;
 import com.pethistory.pet.models.Usuario;
 
 public interface UsuarioMapper {
     Usuario toUsuario(DtoUsuario dtoUsuario);
     DtoUsuario toDtoUsuario(Usuario usuario);
+    DtoUsuarioGett toDtoUsuarioGett (Usuario usuario);
 }
