+++ pet/src/main/java/com/pethistory/pet/service/HistoriaclinicaServiceImplement.java
@@ -1,5 +1,37 @@
 package com.pethistory.pet.service;
 
-public class HistoriaclinicaServiceImplement implements HistoriaclinicaService{
+import java.util.List;
+import java.util.stream.Collector;
+import java.util.stream.Collectors;
 
+import com.pethistory.pet.dtos.DtoMascota;
+import com.pethistory.pet.dtos.HistoriaClinicaDto;
+import com.pethistory.pet.mapper.HistoriaClinicaMapper;
+import com.pethistory.pet.models.HistoriaClinica;
+import com.pethistory.pet.repositories.HistoriaClinicaRepositories;
+
+import jakarta.persistence.EntityNotFoundException;
+
+public class HistoriaclinicaServiceImplement implements HistoriaclinicaService{
+    private final HistoriaClinicaRepositories hiclirep;
+    private final HistoriaClinicaMapper hiclimapp;
+
+    public HistoriaclinicaServiceImplement(HistoriaClinicaRepositories hiclirep, HistoriaClinicaMapper hisclimapp){
+        this.hiclirep=hiclirep;
+        this.hiclimapp=hisclimapp;
+    }
+    @Override
+    public HistoriaClinicaDto crear (HistoriaClinicaDto historiaClinicaDto){
+        HistoriaClinica historiaClinica=hiclimapp.tHistoriaClinica(historiaClinicaDto);
+        HistoriaClinica savee = hiclirep.save(historiaClinica);
+        return hiclimapp.toHistoriaClinicaDto(savee);
+    }
+    @Override
+    public HistoriaClinicaDto buscarId(Long id){
+        return hiclirep.findById(id).map(hiclimapp::toHistoriaClinicaDto).orElseThrow(()->new EntityNotFoundException("historia no encontrada"));
+    }
+    @Override
+    public List<HistoriaClinicaDto> todaslasHis(){
+        return hiclirep.findAll().stream().map(hiclimapp::toHistoriaClinicaDto).collect(Collectors.toList());
+    }
 }
