--- pet/src/main/java/com/pethistory/pet/service/UsuarioGettService.java
@@ -1,7 +1,0 @@
-package com.pethistory.pet.service;
-
-import com.pethistory.pet.dtos.DtoUsuarioGett;
-
-public interface UsuarioGettService {
-    DtoUsuarioGett buscarId(Long id);
-}
